feat: planning-profile drafts, industry-template handlers, calendar-selection official-override
Adds pending planning-profile draft persistence (parameter catalog, validator,
draft repository/store), industry-template handlers/DTOs/queries, a setting
value validator for ACTIVE_INDUSTRIES-style keys, and calendar-selection
official-override support (new SelectedCalendar field + migration). Also
carries scheduling-rule preset/repository changes, region-setup/deploy JSON
updates across ~20 country configs, and language-plugin synonym refreshes.

New migrations: AddPendingPlanningProfileDraft, AddDecimalMinRestDays,
AddSelectedCalendarOfficialOverride.

Work from a concurrently active session on this same tree; committed as a
checkpoint of its current (build-green, test-green) state.

+++ Application/DTOs/Scheduling/SchedulingRuleResource.cs
@@ -10,7 +10,7 @@ public class SchedulingRuleResource
 
     public int? MaxWorkDays { get; set; }
 
-    public int? MinRestDays { get; set; }
+    public decimal? MinRestDays { get; set; }
 
     public decimal? MinPauseHours { get; set; }
