+++ Application/DTOs/Setup/RegionSetupSchedulingRulePreset.cs
@@ -17,7 +17,7 @@ public class RegionSetupSchedulingRulePreset
 
     public int? MaxWorkDays { get; set; }
 
-    public int? MinRestDays { get; set; }
+    public decimal? MinRestDays { get; set; }
 
     public decimal? MinPauseHours { get; set; }
